Refectored the source into a more well-structured structure.
